requirements.txt
pandas
numpy
sklearn
matplotlib
flask
gunicorn
sklearn
PyYAML
evidently
dill
matplotlib
ipykernel
-e .